requirements.txt

python-telegram-bot
requests
replit
